server/requirements.txt
redis
ffmpeg-python